requirements.txt
fastapi
uvicorn
pillow
opencv-python-headless
ultralytics
python-multipart
